renamed variantSummary, newVariantButton,variantSummaryWebPrev so it is easier to read
